third_party/blink/web_tests/platform/generic/media/W3C/video/currentSrc/currentSrc_nonempty_after_adding_source_child-expected.txt

videoElement.currentSrc should not be empty after adding source element (even with a bogus value)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS true is true
PASS false is false
PASS successfullyParsed is true

TEST COMPLETE


spec reference
